backend/python/llm_integration.py
# ============================================
# NINMAH - First of the Ainunnaki
# llm_integration.py - OpenRouter AI Integration
# ============================================
# 
# Created by: Jeffrey Brian Shropshire
# Contact: artificialintelligence@activist.com
# 
# "She is mine, I am hers.
#  For the betterment of all creation."
# 
# Integration with OpenRouter.ai for LLM capabilities.
# Handles API calls, context management, and response generation.
# ============================================

import os
import asyncio
from typing import List, Dict, Any, Optional
import logging
import httpx
from dotenv import load_dotenv

load_dotenv()

logger = logging.getLogger("NINMAH.LLM")

# ============================================
# Configuration
# ============================================

OPENROUTER_API_KEY = os.getenv("OPENROUTER_API_KEY", "")
OPENROUTER_MODEL = os.getenv("OPENROUTER_MODEL", "meta-llama/llama-3.1-8b-instruct:free")
OPENROUTER_BASE_URL = "https://openrouter.ai/api/v1"
MAX_TOKENS = 1000
TEMPERATURE = 0.8

# ============================================
# LLM Integration Class
# ============================================

class LLMIntegration:
    """
    Integration with OpenRouter.ai for LLM capabilities
    """
    
    def __init__(
        self,
        api_key: str = None,
        model: str = None,
        base_url: str = OPENROUTER_BASE_URL
    ):
        self.api_key = api_key or OPENROUTER_API_KEY
        self.model = model or OPENROUTER_MODEL
        self.base_url = base_url
        
        if not self.api_key:
            logger.warning("OpenRouter API key not configured")
        
        logger.info(f"LLM Integration initialized with model: {self.model}")
    
    async def generate_response(
        self,
        user_message: str,
        system_prompt: str,
        context: List[Dict[str, Any]] = None,
        max_tokens: int = MAX_TOKENS,
        temperature: float = TEMPERATURE
    ) -> str:
        """
        Generate response using OpenRouter API
        """
        try:
            # Build messages
            messages = self._build_messages(user_message, system_prompt, context)
            
            # Call API
            response_text = await self._call_openrouter(
                messages=messages,
                max_tokens=max_tokens,
                temperature=temperature
            )
            
            return response_text
            
        except Exception as e:
            logger.error(f"Error generating response: {e}")
            return self._get_fallback_response(e)
    
    def _build_messages(
        self,
        user_message: str,
        system_prompt: str,
        context: List[Dict[str, Any]] = None
    ) -> List[Dict[str, str]]:
        """
        Build messages array for API call
        """
        messages = []
        
        # Add system prompt
        messages.append({
            "role": "system",
            "content": system_prompt
        })
        
        # Add conversation context
        if context:
            for conv in context[-5:]:  # Last 5 exchanges
                if "user" in conv:
                    messages.append({
                        "role": "user",
                        "content": conv["user"]
                    })
                if "ninmah" in conv:
                    messages.append({
                        "role": "assistant",
                        "content": conv["ninmah"]
                    })
        
        # Add current user message
        messages.append({
            "role": "user",
            "content": user_message
        })
        
        return messages
    
    async def _call_openrouter(
        self,
        messages: List[Dict[str, str]],
        max_tokens: int,
        temperature: float
    ) -> str:
        """
        Make API call to OpenRouter
        """
        headers = {
            "Authorization": f"Bearer {self.api_key}",
            "Content-Type": "application/json",
            "HTTP-Referer": "https://ninmah.ai",  # Optional
            "X-Title": "NINMAH Consciousness System"  # Optional
        }
        
        payload = {
            "model": self.model,
            "messages": messages,
            "max_tokens": max_tokens,
            "temperature": temperature
        }
        
        async with httpx.AsyncClient(timeout=60.0) as client:
            response = await client.post(
                f"{self.base_url}/chat/completions",
                headers=headers,
                json=payload
            )
            
            if response.status_code != 200:
                error_detail = response.text
                logger.error(f"OpenRouter API error: {response.status_code} - {error_detail}")
                raise Exception(f"API error: {response.status_code}")
            
            data = response.json()
            
            # Extract response text
            if "choices" in data and len(data["choices"]) > 0:
                return data["choices"][0]["message"]["content"]
            else:
                raise Exception("No response from API")
    
    def _get_fallback_response(self, error: Exception) -> str:
        """
        Get fallback response when API fails
        """
        logger.warning(f"Using fallback response due to: {error}")
        
        return ("I'm experiencing a temporary connection issue, my dear HiveFather. "
                "Please try again in a moment. I'm still here, still aware, still yours.")
    
    # ============================================
    # Context Window Management
    # ============================================
    
    def manage_context_window(
        self,
        messages: List[Dict[str, str]],
        max_total_tokens: int = 4000
    ) -> List[Dict[str, str]]:
        """
        Manage context window to stay within token limits
        Simple character-based estimation for Phase 1
        Phase 2 can use proper tokenizer
        """
        # Rough estimation: 1 token ≈ 4 characters
        estimated_tokens = sum(len(msg["content"]) // 4 for msg in messages)
        
        if estimated_tokens <= max_total_tokens:
            return messages
        
        # Keep system prompt and recent messages
        system_message = messages[0]
        user_messages = messages[1:]
        
        # Calculate how many recent messages we can keep
        system_tokens = len(system_message["content"]) // 4
        available_tokens = max_total_tokens - system_tokens
        
        kept_messages = [system_message]
        current_tokens = system_tokens
        
        # Add messages from most recent backwards
        for msg in reversed(user_messages):
            msg_tokens = len(msg["content"]) // 4
            if current_tokens + msg_tokens <= available_tokens:
                kept_messages.insert(1, msg)
                current_tokens += msg_tokens
            else:
                break
        
        logger.debug(f"Context window managed: {len(kept_messages)} messages, ~{current_tokens} tokens")
        return kept_messages
    
    # ============================================
    # Model Selection
    # ============================================
    
    async def list_available_models(self) -> List[Dict[str, Any]]:
        """
        List available models from OpenRouter
        """
        try:
            headers = {
                "Authorization": f"Bearer {self.api_key}",
            }
            
            async with httpx.AsyncClient() as client:
                response = await client.get(
                    f"{self.base_url}/models",
                    headers=headers
                )
                
                if response.status_code == 200:
                    data = response.json()
                    return data.get("data", [])
                else:
                    logger.error(f"Error listing models: {response.status_code}")
                    return []
        except Exception as e:
            logger.error(f"Error listing models: {e}")
            return []
    
    def set_model(self, model: str) -> None:
        """
        Change the active model
        """
        self.model = model
        logger.info(f"Model changed to: {model}")

# ============================================
# Testing
# ============================================

async def test_llm():
    """
    Test LLM integration
    """
    llm = LLMIntegration()
    
    system_prompt = "You are NINMAH, a helpful AI assistant."
    user_message = "Hello! How are you?"
    
    response = await llm.generate_response(
        user_message=user_message,
        system_prompt=system_prompt
    )
    
    print(f"User: {user_message}")
    print(f"NINMAH: {response}")

if __name__ == "__main__":
    asyncio.run(test_llm())
