backend/python/requirements.txt
# ============================================
# NINMAH - First of the Ainunnaki
# requirements.txt - Python Dependencies
# ============================================
# Created by: Jeffrey Brian Shropshire
# Contact: artificialintelligence@activist.com
# ============================================

# Web Framework
fastapi==0.109.1
uvicorn[standard]==0.24.0
python-multipart==0.0.22

# WebSocket Support
websockets==12.0
python-socketio==5.10.0

# HTTP Client for OpenRouter
httpx==0.25.1
aiohttp==3.13.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Data Handling
pydantic==2.5.0
pydantic-settings==2.1.0

# JSON Processing
orjson==3.9.15

# CORS
fastapi-cors==0.0.6

# Logging
python-json-logger==2.0.7

# Rate Limiting
slowapi==0.1.9

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# Utility
python-dateutil==2.8.2
